fix: repair pre-commit config and add FSM config wrapper
