+++ docs/utils_function_catalog.md
@@ -1,0 +1,160 @@
+# Utils Function Catalog
+
+Summary of top-level functions and classes within directories containing 'utils'.
+
+## src/fsm/utils/config.py
+
+## src/utils/__init__.py
+- Functions: main
+
+## src/utils/caching.py
+- Functions: calculate_file_hash, generate_cache_key
+
+## src/utils/cli_utils.py
+- Functions: main, run_smoke_test
+- Classes: CLIExecutor
+
+## src/utils/config_core/__init__.py
+
+## src/utils/config_core/config_loader.py
+- Classes: ConfigLoader
+
+## src/utils/config_core/config_manager.py
+- Classes: UnifiedConfigManager
+
+## src/utils/config_core/config_validator.py
+- Classes: ConfigValidator
+
+## src/utils/config_core/environment_manager.py
+- Classes: EnvironmentManager
+
+## src/utils/config_core/fsm_config.py
+- Classes: FSMConfig
+
+## src/utils/config_core/unified_configuration_system.py
+- Classes: UnifiedConfigurationSystem
+
+## src/utils/dependency_checker.py
+- Functions: main, run_smoke_test
+- Classes: DependencyChecker
+
+## src/utils/environment_overrides.py
+- Functions: main, run_smoke_test
+- Classes: EnvironmentOverrides
+
+## src/utils/file_utils.py
+- Classes: FileUtils
+
+## src/utils/learning_utils.py
+- Functions: format_learning_result, get_learning_metrics, validate_learning_config
+
+## src/utils/logging_core/__init__.py
+
+## src/utils/logging_core/logging_config.py
+- Classes: LoggingConfig
+
+## src/utils/logging_core/logging_manager.py
+- Classes: UnifiedLoggingManager
+
+## src/utils/logging_core/logging_setup.py
+- Classes: LoggingSetup
+
+## src/utils/logging_core/unified_logging_system.py
+- Classes: UnifiedLoggingSystem
+
+## src/utils/math_utils.py
+- Functions: calculate_mean, clamp
+
+## src/utils/profiling.py
+- Functions: time_block
+
+## src/utils/serializable.py
+- Classes: SerializableMixin
+
+## src/utils/stability_improvements.py
+- Functions: cleanup_stability_improvements, safe_import, setup_stability_improvements, stable_function_call, suppress_warnings_context, validate_imports
+- Classes: StabilityManager
+
+## src/utils/string_utils.py
+- Functions: format_response, generate_hash, get_current_timestamp
+
+## src/utils/system_info.py
+- Functions: main, run_smoke_test
+- Classes: SystemInfo
+
+## src/utils/system_utils_coordinator.py
+- Functions: main, run_smoke_test
+- Classes: SystemUtilsCoordinator
+
+## src/utils/test_complete_utility_system.py
+- Functions: main, print_error, print_header, print_info, print_section, print_success, print_warning
+
+## src/utils/test_logging_consolidation.py
+- Functions: test_complete_consolidation, test_duplicate_elimination, test_logging_consolidation
+
+## src/utils/test_validation_consolidation.py
+- Functions: test_duplicate_elimination, test_validation_consolidation
+
+## src/utils/validation_core/__init__.py
+
+## src/utils/validation_core/base_validator.py
+- Classes: BaseValidator
+
+## src/utils/validation_core/collection_validators.py
+- Classes: CollectionValidators
+
+## src/utils/validation_core/data_validators.py
+- Classes: DataValidators
+
+## src/utils/validation_core/date_validators.py
+- Classes: DateValidators
+
+## src/utils/validation_core/email_validators.py
+- Classes: EmailValidators
+
+## src/utils/validation_core/format_validators.py
+- Classes: FormatValidators
+
+## src/utils/validation_core/numeric_validators.py
+- Classes: NumericValidators
+
+## src/utils/validation_core/performance_tracker.py
+- Classes: PerformanceTracker
+
+## src/utils/validation_core/string_validators.py
+- Classes: StringValidators
+
+## src/utils/validation_core/unified_validation_system.py
+- Classes: UnifiedValidationSystem
+
+## src/utils/validation_core/url_validators.py
+- Classes: URLValidators
+
+## src/utils/validation_core/validation_result.py
+- Classes: ValidationResult, ValidationStatus
+
+## src/utils/validation_core/value_validators.py
+- Classes: ValueValidators
+
+## tests/utils/__init__.py
+
+## tests/utils/api_keys.py
+- Functions: set_api_keys
+
+## tests/utils/mock_managers.py
+- Classes: MockAgentManager, MockFSMOrchestrator, MockResponseCaptureService, MockTaskManager, MockWorkflowEngine, MockWorkspaceManager
+
+## tests/utils/test_config_manager.py
+- Classes: TestConfigManager
+
+## tests/utils/test_data.py
+- Functions: get_api_test_data, get_integration_test_data, get_performance_test_data, get_sample_agent_data, get_sample_config_data, get_sample_task_data, get_security_test_data
+
+## tests/utils/test_file_utils.py
+- Classes: TestFileUtils
+
+## tests/utils/test_system_utils.py
+- Classes: TestSystemUtils
+
+## tests/utils/test_validation_utils.py
+- Classes: TestValidationUtils
