+++ src/ai_ml/utilities/common_utils.py
@@ -1,21 +1,10 @@
-"""
-Common Utility Functions
-Captain Agent-3: Shared, Non-Duplicated Utilities
-"""
+"""Common AI/ML utilities leveraging centralized implementations."""
 
-import json
-import hashlib
 from typing import Any, Dict
+from src.utils.string_utils import generate_hash, format_response
 
-def generate_hash(data: Any) -> str:
-    """Generate hash for data - single implementation"""
-    return hashlib.md5(json.dumps(data, sort_keys=True).encode()).hexdigest()
 
-def validate_config(config: Dict) -> bool:
+def validate_config(config: Dict[str, Any]) -> bool:
     """Validate configuration - single implementation"""
-    required_keys = ['name', 'version', 'status']
+    required_keys = ["name", "version", "status"]
     return all(key in config for key in required_keys)
-
-def format_response(data: Any, status: str = "success") -> Dict:
-    """Format response - single implementation"""
-    return {"status": status, "data": data, "timestamp": "2025-08-28T22:15:00.000000Z"}
