--- src/core/utils/consolidated_utils.py
@@ -1,128 +1,0 @@
-#!/usr/bin/env python3
-"""
-Consolidated utils.py - Agent-8 Autonomous Cleanup
-================================================
-
-Automatically consolidated utils.py from all scattered locations.
-Generated by Agent-8 during autonomous cleanup mission.
-
-Author: Agent-8 (Integration Enhancement Optimization Manager)
-Mission: Autonomous Cleanup & Side Mission Completion
-Status: MAXIMUM EFFICIENCY MODE
-"""
-
-from typing import Dict, Any, List, Optional
-from dataclasses import dataclass
-from enum import Enum
-import os
-import logging
-
-# ============================================================================
-# WORKFLOW UTILS
-# ============================================================================
-
-# Consolidated from .\src\core\workflow\utils.py
-# Original file: .\src\core\workflow\utils.py
-# Category: workflow
-# TODO: Import or implement functionality from .\src\core\workflow\utils.py
-# TODO: Remove duplicate file after migration
-
-# Consolidated from .\src\core\refactoring\automated_workflow_orchestrator\utils.py
-# Original file: .\src\core\refactoring\automated_workflow_orchestrator\utils.py
-# Category: workflow
-# TODO: Import or implement functionality from .\src\core\refactoring\automated_workflow_orchestrator\utils.py
-# TODO: Remove duplicate file after migration
-
-# Consolidated from .\src\core\refactoring\workflow_reliability_testing\utils.py
-# Original file: .\src\core\refactoring\workflow_reliability_testing\utils.py
-# Category: workflow
-# TODO: Import or implement functionality from .\src\core\refactoring\workflow_reliability_testing\utils.py
-# TODO: Remove duplicate file after migration
-
-# ============================================================================
-# FSM UTILS
-# ============================================================================
-
-# Consolidated from .\src\core\fsm\utils.py
-# Original file: .\src\core\fsm\utils.py
-# Category: fsm
-# TODO: Import or implement functionality from .\src\core\fsm\utils.py
-# TODO: Remove duplicate file after migration
-
-# ============================================================================
-# HANDOFF UTILS
-# ============================================================================
-
-# Consolidated from .\src\core\handoff_reliability\utils.py
-# Original file: .\src\core\handoff_reliability\utils.py
-# Category: handoff
-# TODO: Import or implement functionality from .\src\core\handoff_reliability\utils.py
-# TODO: Remove duplicate file after migration
-
-# ============================================================================
-# AGENT MANAGEMENT UTILS
-# ============================================================================
-
-# Consolidated from .\src\core\agent\utils.py
-# Original file: .\src\core\agent\utils.py
-# Category: agent_management
-# TODO: Import or implement functionality from .\src\core\agent\utils.py
-# TODO: Remove duplicate file after migration
-
-# ============================================================================
-# COMMUNICATION UTILS
-# ============================================================================
-
-# Consolidated from .\src\services\communication\utils.py
-# Original file: .\src\services\communication\utils.py
-# Category: communication
-# TODO: Import or implement functionality from .\src\services\communication\utils.py
-# TODO: Remove duplicate file after migration
-
-# Consolidated from .\src\communications\utils.py
-# Original file: .\src\communications\utils.py
-# Category: communication
-# TODO: Import or implement functionality from .\src\communications\utils.py
-# TODO: Remove duplicate file after migration
-
-# ============================================================================
-# WEB FRONTEND UTILS
-# ============================================================================
-
-# Consolidated from .\src\web\frontend\utils.py
-# Original file: .\src\web\frontend\utils.py
-# Category: web_frontend
-# TODO: Import or implement functionality from .\src\web\frontend\utils.py
-# TODO: Remove duplicate file after migration
-
-# ============================================================================
-# TESTING UTILS
-# ============================================================================
-
-# Consolidated from .\src\testing\utils.py
-# Original file: .\src\testing\utils.py
-# Category: testing
-# TODO: Import or implement functionality from .\src\testing\utils.py
-# TODO: Remove duplicate file after migration
-
-# ============================================================================
-# REFACTORING UTILS
-# ============================================================================
-
-# Consolidated from .\refactoring_performance_benchmark\utils.py
-# Original file: .\refactoring_performance_benchmark\utils.py
-# Category: refactoring
-# TODO: Import or implement functionality from .\refactoring_performance_benchmark\utils.py
-# TODO: Remove duplicate file after migration
-
-# ============================================================================
-# CONSOLIDATION METADATA
-# ============================================================================
-
-CONSOLIDATION_INFO = {
-    "pattern": "utils.py",
-    "total_files": 11,
-    "categories": 8,
-    "consolidated_by": "Agent-8",
-    "mission": "Autonomous Cleanup & Side Mission Completion"
-}
