--- src/core/utils/consolidated_utils_system.py
@@ -1,393 +1,0 @@
-"""
-🎯 CONSOLIDATED UTILS SYSTEM - SINGLE SOURCE OF TRUTH
-Agent-7 - Autonomous Cleanup Mission
-
-Consolidated utility functions from scattered implementations.
-Eliminates SSOT violations by providing unified utilities for all systems.
-
-This module consolidates utilities from:
-- src/core/utils/
-- src/utils/
-- Multiple scattered utility implementations
-
-Agent: Agent-7 (Quality Completion Optimization Manager)
-Mission: AUTONOMOUS CLEANUP - Multiple side missions in one cycle
-Priority: CRITICAL - Maximum efficiency
-Status: IMPLEMENTATION PHASE 1 - Unified Utils System
-
-Author: Agent-7 - Quality Completion Optimization Manager
-License: MIT
-"""
-
-import os
-import sys
-import json
-import logging
-import hashlib
-import shutil
-from pathlib import Path
-from typing import Dict, Any, List, Optional, Union
-from datetime import datetime
-from collections import defaultdict
-import subprocess
-import platform
-import psutil
-import threading
-import time
-
-
-class ConsolidatedUtilsSystem:
-    """
-    Unified utility system for all utility functions.
-    
-    Consolidates utility functionality from scattered implementations
-    into a single source of truth.
-    """
-    
-    def __init__(self):
-        """Initialize the consolidated utils system."""
-        self.logger = logging.getLogger(f"{__name__}.ConsolidatedUtilsSystem")
-        self.consolidation_status = {
-            "utils_consolidated": 0,
-            "original_locations": [],
-            "consolidation_status": "IN_PROGRESS",
-            "v2_compliance": "VERIFIED"
-        }
-        
-        # Initialize core utilities
-        self._initialize_core_utilities()
-        
-        self.logger.info("✅ Consolidated Utils System initialized for autonomous cleanup mission")
-    
-    def _initialize_core_utilities(self):
-        """Initialize core utility functions."""
-        # File system utilities
-        self.file_utils = FileSystemUtils()
-        
-        # String utilities
-        self.string_utils = StringUtils()
-        
-        # Math utilities
-        self.math_utils = MathUtils()
-        
-        # IO utilities
-        self.io_utils = IOUtils()
-        
-        # System utilities
-        self.system_utils = SystemUtils()
-        
-        # Validation utilities
-        self.validation_utils = ValidationUtils()
-        
-        # Logging utilities
-        self.logging_utils = LoggingUtils()
-        
-        # Configuration utilities
-        self.config_utils = ConfigUtils()
-        
-        self.logger.info(f"✅ Initialized {8} core utility modules")
-    
-    def consolidate_utils_directories(self) -> Dict[str, Any]:
-        """Consolidate scattered utils directories into unified system."""
-        consolidation_results = {
-            "directories_consolidated": 0,
-            "files_consolidated": 0,
-            "duplicates_removed": 0,
-            "errors": []
-        }
-        
-        try:
-            # Identify utils directories
-            utils_directories = [
-                "src/core/utils",
-                "src/utils",
-                "src/core/validation",
-                "src/core/configuration"
-            ]
-            
-            for directory in utils_directories:
-                if os.path.exists(directory):
-                    consolidation_results["directories_consolidated"] += 1
-                    consolidation_results["files_consolidated"] += self._consolidate_directory(directory)
-            
-            self.logger.info(f"✅ Consolidated {consolidation_results['directories_consolidated']} utils directories")
-            return consolidation_results
-            
-        except Exception as e:
-            error_msg = f"Error consolidating utils directories: {e}"
-            consolidation_results["errors"].append(error_msg)
-            self.logger.error(f"❌ {error_msg}")
-            return consolidation_results
-    
-    def _consolidate_directory(self, directory: str) -> int:
-        """Consolidate a single directory into unified system."""
-        files_consolidated = 0
-        
-        try:
-            for root, dirs, files in os.walk(directory):
-                for file in files:
-                    if file.endswith('.py'):
-                        source_path = os.path.join(root, file)
-                        target_path = self._get_consolidated_path(source_path)
-                        
-                        if self._should_consolidate_file(source_path, target_path):
-                            self._consolidate_file(source_path, target_path)
-                            files_consolidated += 1
-                            
-        except Exception as e:
-            self.logger.error(f"Error consolidating directory {directory}: {e}")
-        
-        return files_consolidated
-    
-    def _get_consolidated_path(self, source_path: str) -> str:
-        """Get the consolidated path for a source file."""
-        # Map source paths to consolidated structure
-        path_mapping = {
-            "src/core/utils": "src/core/utils/consolidated",
-            "src/utils": "src/core/utils/consolidated",
-            "src/core/validation": "src/core/utils/consolidated/validation",
-            "src/core/configuration": "src/core/utils/consolidated/config"
-        }
-        
-        for source_dir, target_dir in path_mapping.items():
-            if source_path.startswith(source_dir):
-                relative_path = os.path.relpath(source_path, source_dir)
-                return os.path.join(target_dir, relative_path)
-        
-        return source_path
-    
-    def _should_consolidate_file(self, source_path: str, target_path: str) -> bool:
-        """Determine if a file should be consolidated."""
-        # Skip if target already exists and is newer
-        if os.path.exists(target_path):
-            source_time = os.path.getmtime(source_path)
-            target_time = os.path.getmtime(target_path)
-            if target_time >= source_time:
-                return False
-        
-        # Skip backup files
-        if source_path.endswith('.backup'):
-            return False
-        
-        return True
-    
-    def _consolidate_file(self, source_path: str, target_path: str):
-        """Consolidate a single file."""
-        try:
-            # Ensure target directory exists
-            os.makedirs(os.path.dirname(target_path), exist_ok=True)
-            
-            # Copy file to consolidated location
-            shutil.copy2(source_path, target_path)
-            
-            self.logger.debug(f"✅ Consolidated: {source_path} → {target_path}")
-            
-        except Exception as e:
-            self.logger.error(f"Error consolidating file {source_path}: {e}")
-    
-    def get_consolidation_status(self) -> Dict[str, Any]:
-        """Get overall consolidation status."""
-        return {
-            "system_name": "Consolidated Utils System",
-            "consolidation_status": self.consolidation_status,
-            "core_modules": [
-                "FileSystemUtils",
-                "StringUtils", 
-                "MathUtils",
-                "IOUtils",
-                "SystemUtils",
-                "ValidationUtils",
-                "LoggingUtils",
-                "ConfigUtils"
-            ],
-            "v2_compliance": "VERIFIED",
-            "ssot_compliance": "ACHIEVED"
-        }
-
-
-class FileSystemUtils:
-    """Unified file system utilities."""
-    
-    def ensure_directory(self, path: str) -> bool:
-        """Ensure directory exists, create if needed."""
-        try:
-            os.makedirs(path, exist_ok=True)
-            return True
-        except Exception:
-            return False
-    
-    def safe_remove(self, path: str) -> bool:
-        """Safely remove file or directory."""
-        try:
-            if os.path.isfile(path):
-                os.remove(path)
-            elif os.path.isdir(path):
-                shutil.rmtree(path)
-            return True
-        except Exception:
-            return False
-    
-    def get_file_hash(self, file_path: str) -> Optional[str]:
-        """Get SHA256 hash of file."""
-        try:
-            with open(file_path, 'rb') as f:
-                return hashlib.sha256(f.read()).hexdigest()
-        except Exception:
-            return None
-
-
-class StringUtils:
-    """Unified string utilities."""
-    
-    def sanitize_filename(self, filename: str) -> str:
-        """Sanitize filename for safe file system use."""
-        import re
-        # Remove or replace unsafe characters
-        sanitized = re.sub(r'[<>:"/\\|?*]', '_', filename)
-        return sanitized.strip()
-    
-    def truncate_with_ellipsis(self, text: str, max_length: int) -> str:
-        """Truncate text with ellipsis if too long."""
-        if len(text) <= max_length:
-            return text
-        return text[:max_length-3] + "..."
-    
-    def normalize_whitespace(self, text: str) -> str:
-        """Normalize whitespace in text."""
-        import re
-        return re.sub(r'\s+', ' ', text).strip()
-
-
-class MathUtils:
-    """Unified math utilities."""
-    
-    def calculate_percentage(self, part: float, total: float) -> float:
-        """Calculate percentage."""
-        if total == 0:
-            return 0.0
-        return (part / total) * 100.0
-    
-    def round_to_decimal(self, value: float, decimal_places: int) -> float:
-        """Round value to specified decimal places."""
-        return round(value, decimal_places)
-    
-    def clamp_value(self, value: float, min_val: float, max_val: float) -> float:
-        """Clamp value between min and max."""
-        return max(min_val, min(value, max_val))
-
-
-class IOUtils:
-    """Unified IO utilities."""
-    
-    def read_json_file(self, file_path: str) -> Optional[Dict[str, Any]]:
-        """Read JSON file safely."""
-        try:
-            with open(file_path, 'r', encoding='utf-8') as f:
-                return json.load(f)
-        except Exception:
-            return None
-    
-    def write_json_file(self, file_path: str, data: Dict[str, Any]) -> bool:
-        """Write JSON file safely."""
-        try:
-            with open(file_path, 'w', encoding='utf-8') as f:
-                json.dump(data, f, indent=2, ensure_ascii=False)
-            return True
-        except Exception:
-            return False
-    
-    def read_text_file(self, file_path: str) -> Optional[str]:
-        """Read text file safely."""
-        try:
-            with open(file_path, 'r', encoding='utf-8') as f:
-                return f.read()
-        except Exception:
-            return None
-
-
-class SystemUtils:
-    """Unified system utilities."""
-    
-    def get_system_info(self) -> Dict[str, Any]:
-        """Get system information."""
-        return {
-            "platform": platform.system(),
-            "platform_version": platform.version(),
-            "python_version": sys.version,
-            "cpu_count": os.cpu_count(),
-            "memory_usage": psutil.virtual_memory().percent
-        }
-    
-    def get_process_info(self) -> Dict[str, Any]:
-        """Get current process information."""
-        process = psutil.Process()
-        return {
-            "pid": process.pid,
-            "memory_usage": process.memory_info().rss,
-            "cpu_percent": process.cpu_percent()
-        }
-    
-    def is_system_healthy(self) -> bool:
-        """Check if system is healthy."""
-        try:
-            memory = psutil.virtual_memory()
-            return memory.percent < 90  # Less than 90% memory usage
-        except Exception:
-            return False
-
-
-class ValidationUtils:
-    """Unified validation utilities."""
-    
-    def validate_file_exists(self, file_path: str) -> bool:
-        """Validate file exists."""
-        return os.path.isfile(file_path)
-    
-    def validate_directory_exists(self, dir_path: str) -> bool:
-        """Validate directory exists."""
-        return os.path.isdir(dir_path)
-    
-    def validate_json_structure(self, data: Dict[str, Any], required_keys: List[str]) -> bool:
-        """Validate JSON structure has required keys."""
-        return all(key in data for key in required_keys)
-
-
-class LoggingUtils:
-    """Unified logging utilities."""
-    
-    def setup_logging(self, log_level: str = "INFO") -> None:
-        """Setup logging configuration."""
-        logging.basicConfig(
-            level=getattr(logging, log_level.upper()),
-            format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
-            handlers=[
-                logging.StreamHandler(),
-                logging.FileHandler('consolidated_utils.log')
-            ]
-        )
-    
-    def get_logger(self, name: str) -> logging.Logger:
-        """Get logger with specified name."""
-        return logging.getLogger(name)
-
-
-class ConfigUtils:
-    """Unified configuration utilities."""
-    
-    def load_config_file(self, config_path: str) -> Optional[Dict[str, Any]]:
-        """Load configuration file."""
-        return self.read_json_file(config_path)
-    
-    def save_config_file(self, config_path: str, config_data: Dict[str, Any]) -> bool:
-        """Save configuration file."""
-        return self.write_json_file(config_path, config_data)
-    
-    def merge_configs(self, base_config: Dict[str, Any], override_config: Dict[str, Any]) -> Dict[str, Any]:
-        """Merge two configurations, override taking precedence."""
-        merged = base_config.copy()
-        merged.update(override_config)
-        return merged
-
-
-# Global instance for easy access
-consolidated_utils = ConsolidatedUtilsSystem()
